requirements.txt
# Core dependencies (keep existing)
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
lxml>=4.9.0
praw>=7.7.0
google-generativeai>=0.3.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10

# Vector Database
pinecone>=7.0.0

# HuggingFace (for LLM and embeddings inference)
huggingface-hub>=0.20.0
requests>=2.28.0

# Sentiment Analysis (lightweight - no GPU needed)
vaderSentiment>=3.3.2

# Database
sqlalchemy>=2.0.0

# Web UI
streamlit>=1.29.0

# Environment
python-dotenv>=1.0.0
